Create js files for copter, boundary, and obstacles
Use module exports to include files
