+++ lib/boundary.js
@@ -1,0 +1,18 @@
+var canvas = document.getElementById('game');
+var context = canvas.getContext('2d');
+
+class Boundary {
+  constructor(x, y, width, height) {
+    this.x = x;
+    this.y = y;
+    this.width = width;
+    this.height = height;
+  }
+
+  draw() {
+    context.fillRect(this.x, this.y, this.width, this.height);
+    return this;
+  }
+}
+
+module.exports = Boundary;
